simplifying limit enforcement structures (#606)
